+++ src/components/home/sectionSecond/styled.js
