+++ src/page/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Header from '../components/header/Header'
-import MainInstitucional from '../components/main-institucional/MainInstitucional'
 import Footer from '../components/footer/Footer'
+import Home from '../components/home/Home'
 import { Container, GlobalStyle } from './styled'
 
 export default class App extends React.Component {
@@ -10,7 +10,7 @@ export default class App extends React.Component {
 			<Container>
 				<GlobalStyle />
 				<Header/>
-				<MainInstitucional/>
+				<Home />
 				<Footer/>
 			</Container>
 		)
